[Feature] add error pages
